refactor: remove complex service call add some Javadoc, static resource
