+++ profanity-api/src/main/resources/static/index.html
